Update(Design): Delete MenuCategory
Don't show delete option in dropdown menu on MenuCategoryIndex when only one exists.
Prompt to reassign related Recipes to another MenuCategory when they exist.
Create Modal blade component and trigger in app.js

+++ app/Http/Livewire/MenuCategoryIndex.php
@@ -12,6 +12,10 @@ class MenuCategoryIndex extends LivewireBaseComponent
 
     public string $menuCategoryEditing = '';
 
+    public string $wantingToDelete = '';
+
+    public string $categoryIdToMoveTo = '';
+
     protected array $rules = [
         'nameInput' => 'required',
         'costingGoalInput' => 'required|numeric',
@@ -23,6 +27,13 @@ class MenuCategoryIndex extends LivewireBaseComponent
             return 'Create';
         }
         return 'Edit';
+    }
+
+    public function getCategoriesForMoveProperty(): \Illuminate\Support\Collection
+    {
+        $categories = MenuCategory::select(['id', 'name'])->whereNotIn('id', [$this->wantingToDelete])->get();
+        $this->categoryIdToMoveTo = $categories->first()->id ?? '';
+        return $categories;
     }
 
     public function loadMenuCategoryToEdit(MenuCategory $menuCategory): void
@@ -74,8 +85,30 @@ class MenuCategoryIndex extends LivewireBaseComponent
         $this->alertWithToast('Updated.');
     }
 
+    public function moveAllToCategory(): void
+    {
+        $menuCategoryToMoveFrom = MenuCategory::findOrFail($this->wantingToDelete);
+
+        foreach ($menuCategoryToMoveFrom->recipes as $recipe) {
+            $recipe->update([
+                'menu_category_id' => $this->categoryIdToMoveTo
+            ]);
+        }
+
+        $this->alertWithToast('Recipes moved.');
+        $this->deleteMenuCategory($menuCategoryToMoveFrom);
+    }
+
     public function deleteMenuCategory(MenuCategory $menuCategory): void
     {
+        $menuCategory->loadCount('recipes');
+
+        if ($menuCategory->recipes_count != 0) {
+            $this->wantingToDelete = $menuCategory->id;
+            $this->emit('showModal', 'moveModal');
+            return;
+        }
+
         $menuCategory->delete();
         $this->alertWithToast('Deleted.');
     }
